Babel. Revista de Libros_ Force Atlas 2
The graph uses the built-in layout Force Atlas in order to display the edges´ direction of every line between two nodes. It encompasses not only the social threads between the writers who belong to the network but also the external relations these writers made with people beyond the grid. Consequently, the image invites to assess the diversity of external relations some nodes have in comparison to other writers, who only maintain interactions with others inside the circle.